requirements.txt
aiohttp==3.7.4.post0
alembic==1.7.6
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
Mako==1.1.6
MarkupSafe==2.1.0
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
py-cord==2.0.0b4
pyparsing==3.0.7
pytest==7.0.1
PyYAML==6.0
SQLAlchemy==1.4.31
tomli==2.0.1
typing_extensions==4.1.1
yarl==1.7.2
